integrated new hashing classes into MerkleTree
