+++ .vscode/launch.json
@@ -5,7 +5,7 @@
       "name": "Launch",
       "type": "node",
       "request": "launch",
-      "program": "${workspaceRoot}/bin/tests/run2.js",
+      "program": "${workspaceRoot}/bin/tests/run.js",
       "cwd": "${workspaceRoot}/bin",
       "preLaunchTask": "build",
       "env": {
