+++ bin/index.js
@@ -5,6 +5,5 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var MerkleTree_1 = require("./lib/MerkleTree");
 exports.MerkleTree = MerkleTree_1.MerkleTree;
 var hash_1 = require("./lib/hash");
-exports.getHashDigestSize = hash_1.getHashDigestSize;
-exports.getHashFunction = hash_1.getHashFunction;
+exports.createHash = hash_1.createHash;
 //# sourceMappingURL=index.js.map
