+++ index.ts
@@ -1,4 +1,4 @@
 // RE-EXPORTS
 // ================================================================================================
 export { MerkleTree } from './lib/MerkleTree';
-export { getHashDigestSize, getHashFunction } from './lib/hash';
+export { createHash } from './lib/hash';
