+++ lib/MerkleTree.ts
@@ -21,13 +21,11 @@ export class MerkleTree {
 
     static create(values: Buffer[], hashAlgorithm: HashAlgorithm) {
         
-        const buildTree = hashing.getMerkleTreeBuilder(hashAlgorithm);
-        const nodeSize = hashing.getHashDigestSize(hashAlgorithm);
-
+        const hash = hashing.createHash(hashAlgorithm);
         const depth = Math.ceil(Math.log2(values.length));
-        const nodes = buildTree(depth, values)
+        const nodes = hash.buildMerkleNodes(depth, values)
 
-        return new MerkleTree(nodes, values, depth, nodeSize);
+        return new MerkleTree(nodes, values, depth, hash.digestSize);
     }
 
     private constructor(nodes: ArrayBuffer, values: Buffer[], depth: number, nodeSize: number) {
@@ -134,20 +132,20 @@ export class MerkleTree {
     // STATIC METHODS
     // --------------------------------------------------------------------------------------------
     static verify(root: Buffer, index: number, proof: Buffer[], hashAlgorithm: HashAlgorithm): boolean {
-        const hash = hashing.getHashFunction(hashAlgorithm);
+        const hash = hashing.createHash(hashAlgorithm);
 
         const r = index & 1;
         const value1 = proof[r];
         const value2 = proof[1 - r];
-        let v = hash(value1, value2);
+        let v = hash.merge(value1, value2);
 
         index = (index + 2 ** (proof.length - 1)) >> 1;
         for ( let i = 2; i < proof.length; i++) {
             if (index & 1) {
-                v = hash(proof[i], v);
+                v = hash.merge(proof[i], v);
             }
             else {
-                v = hash(v, proof[i]);
+                v = hash.merge(v, proof[i]);
             }
             index = index >> 1;
         }
@@ -157,7 +155,7 @@ export class MerkleTree {
 
     static verifyBatch(root: Buffer, indexes: number[], proof: BatchMerkleProof, hashAlgorithm: HashAlgorithm): boolean {
         const v = new Map<number,Buffer>();
-        const hash = hashing.getHashFunction(hashAlgorithm);
+        const hash = hashing.createHash(hashAlgorithm);
 
         // replace odd indexes, offset, and sort in ascending order
         const offset = 2 ** proof.depth;
@@ -194,7 +192,7 @@ export class MerkleTree {
                 proofPointers[i] = 1;
             }
 
-            let parent = hash(v1, v2);
+            let parent = hash.merge(v1, v2);
             let parentIndex = (offset + index >> 1);
             
             v.set(parentIndex, parent);
@@ -228,7 +226,7 @@ export class MerkleTree {
                 if (node === undefined || sibling === undefined) return false;
 
                 // calculate parent node and add it to the next set of nodes
-                let parent = (nodeIndex & 1) ? hash(sibling, node) : hash(node, sibling);
+                let parent = (nodeIndex & 1) ? hash.merge(sibling, node) : hash.merge(node, sibling);
                 let parentIndex = nodeIndex >> 1;
                 v.set(parentIndex, parent);
                 nextIndexes.push(parentIndex);
